changed type of price, discountPrice and discountPercentage as double

+++ src/main/kotlin/com/example/data/database/table/ProductTable.kt
@@ -6,7 +6,7 @@ import org.jetbrains.exposed.sql.ReferenceOption
 object ProductTable : IntIdTable("products") {
     val name = text("name").uniqueIndex()
     val description = text("description")
-    val price = float("price")
+    val price = double("price")
     val region = text("region")
     val weight = integer("weight")
     val flavorProfile = text("flavor_profile")
@@ -16,7 +16,7 @@ object ProductTable : IntIdTable("products") {
     val category = reference("category", refColumn = CategoryTable.name, onDelete = ReferenceOption.CASCADE)
     val stockQuantity = integer("stock_quantity")
     val popularityRating = float("popularity_rating")
-    val discountPrice = float("discount_price").nullable()
-    val discountPercentage = float("discount_percentage").nullable()
+    val discountPrice = double("discount_price").nullable()
+    val discountPercentage = double("discount_percentage").nullable()
     val bestseller = bool("bestseller")
 }
